semana 2 ejer2Py

+++ Semana1_C#/Program.cs
@@ -29,7 +29,7 @@ namespace Semana1_C_
             int x = int.Parse( Console.ReadLine());
             Console.WriteLine("Ingrese numero y: ");
             int y = Convert.ToInt32( Console.ReadLine());
-            double resu = x / y;
+            double resu = (double)x / (double)y;
             Console.WriteLine("Suma: " + (x + y));
             Console.WriteLine("Resta: " + (x - y));
             Console.WriteLine("Multiplicacion: " + (x * y));
